fix: OP_PUSHDATA4 emits a 4-byte length, not 8
ScriptChunkHelper.scriptData appended UInt64(data.count) (8 bytes) for
OP_PUSHDATA4, where consensus requires a 4-byte little-endian length —
malforming every push of 65536..4GB (large OP_RETURN / inscriptions).
The author reached for UInt64 because UInt32 already conforms to
DataConvertable (4-byte LE) but that wasn't obvious; use UInt32. Add a
regression test asserting the 5-byte prefix (forced + auto encodings).

+++ Sources/SwiftBSV/Scripts/ScriptChunkHelper.swift
@@ -46,7 +46,11 @@ public struct ScriptChunkHelper {
             scriptData += UInt16(data.count)
         } else if UInt64(data.count) <= 0xffffffff && (preferredLengthEncoding == -1 || preferredLengthEncoding == 4) {
             scriptData += OpCode.OP_PUSHDATA4
-            scriptData += UInt64(data.count)
+            // OP_PUSHDATA4 takes a 4-byte little-endian length. This was
+            // UInt64 (8 bytes) — a malformed prefix — because UInt32's
+            // BinaryConvertible conformance was commented out; it is now
+            // enabled in Data+Script.swift.
+            scriptData += UInt32(data.count)
         } else {
             // Invalid preferredLength encoding or data size is too big.
             return nil
